+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 
 ### Fixes
 
+- Ensure width and depth are never attached to `DryAdvection` source when running pointwise-only model.
+- Ensure that azimuth is not dropped when present in `DryAdvection` input source.
 - Exclude netCDF version 1.7.0, which causes non-deterministic failures in unit tests primarily due to segmentation faults.
 - Pin numpy to 1.x in runtime dependencies until a working numpy 2.x-compatible netCDF4 release is available.